[bootstrap] Move `sanitize_sh` from `dist` to `install`
This is the only place it's used, so there's no need for it to be public in another module.
In general, `dist` shouldn't ever touch shell scripts.
